added kms_test.go
